requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pytest==8.3.0
httpx==0.27.0
